Moving vmem to a separate mod
